# Frontend/ops-console/openapi.yaml
openapi: 3.0.3
info:
  title: Vayva Ops Console API
  description: |-
    Internal operations platform API for managing merchants, orders, KYC, disputes, and platform administration.

    Auto-generated from Next.js route files.
  version: 1.0.0
  contact:
    name: Vayva Engineering
    email: engineering@vayva.co
servers:
  - url: https://ops.vayva.co/api/ops
    description: Production server
  - url: http://localhost:3002/api/ops
    description: Development server
security:
  - cookieAuth: []
...
paths:
...
  /industries:
    get:
      summary: List industries
      tags:
        - Industries
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ApiMeta:
      type: object
      properties:
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - requestId
        - timestamp
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          enum:
            - false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
        meta:
          $ref: '#/components/schemas/ApiMeta'
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: vayva_ops_session